+++ docs/changelog.md
@@ -2,3 +2,7 @@
 
 * Search, follow, and unfollow  for tv creators
 * Display latest credited work for followed creators
+
+### Week 4
+
+* Add shows to watch list